requirements.txt
hamolicious-vector-class==0.4.10
numpy==1.21.0
pygame==2.0.1
